kymoo 님이 변경한 내용

+++ src/main/java/com/sensingbros/recommendation/domain/Review.java
@@ -6,6 +6,8 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import lombok.Getter;
 import lombok.Setter;
@@ -21,7 +23,7 @@ public class Review {
     private Integer id;
 
     @Column
-    private Integer score;
+    private BigDecimal score;
 
     @Column(columnDefinition = "text")
     private String body;
